requirements.txt
google-genai
python-dotenv~=1.2.2
protobuf~=5.29.6